python/requirement.txt
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
idna==3.6
joblib==1.3.2
mysql-connector-python==8.2.0
nltk==3.8.1
numpy==1.26.3
oauthlib==3.2.2
pandas==2.1.4
protobuf==4.21.12
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
Sastrawi==1.0.1
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
threadpoolctl==3.2.0
tqdm==4.66.1
tweepy==4.14.0
tzdata==2023.4
urllib3==2.1.0
